backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["timofejkazakov <t.kazakov@agora.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
pillow = "^10.3.0"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.6.0"
redis = "^5.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
